fixed updating prices and added launcher icon

+++ app/src/main/java/xyz/tgprojects/buildmeabudget/models/Category.java
@@ -32,7 +32,7 @@ public class Category implements Serializable{
         this.description = description;
     }
 
-    public long getPercentage() {
+    public int getPercentage() {
         return percentage;
     }
